+++ inst/examples/demo/adam/mk100adsl.R
@@ -1,3 +1,0 @@
-# This script produces no errors or warnings for testing purposes
-
-message("this script has no errors or warnings")