+++ inst/examples/sequence_exuction/test.py
@@ -1,8 +1,0 @@
-1+1
-
-file_path = 'from_python.txt'
-
-with open(file_path, 'w') as file:
-    # Write or save text to the file
-    file.write("Hello from python ^^")
-