+++ inst/examples/simple/sleep_2.R
@@ -1,2 +1,0 @@
-# Dummys cript to test parrallisation of callr
-Sys.sleep(2)