+++ NEWS.md
@@ -1,3 +1,7 @@
+# whirl dev
+* Added support for logging of Python scripts with `run()`.
+* Improved unit tests for `run()`
+
 # whirl 0.1.4 (2024-11-01)
 * Add `use_whirl()` utility function.